--- a/internal/protocol/server_parameters.go
+++ b/internal/protocol/server_parameters.go
@@ -94,7 +94,7 @@ const MaxTrackedReceivedAckRanges = DefaultMaxCongestionWindow
 const MaxNonRetransmittablePackets = 19
 
 // RetransmittablePacketsBeforeAck is the number of retransmittable that an ACK is sent for
-const RetransmittablePacketsBeforeAck = 2
+const RetransmittablePacketsBeforeAck = 10
 
 // MaxStreamFrameSorterGaps is the maximum number of gaps between received StreamFrames
 // prevents DoS attacks against the streamFrameSorter