Ian told me in Melbourne that Chrome is currently running an experiment to send less ACKs, and has already reduced the frequency (from every 2 packets to 10 packets now).
I ran some benchmarks and I'm seeing significant improvements in quic-go as well.

before:
```
+-----------------+----------------------------+--------------------------+--------------------------+
|                 | Chrome => TCP/HTTPS server | Chrome => quic-go server | quic-go client => server |
+-----------------+----------------------------+--------------------------+--------------------------+
| direct transfer | 142.80 ± 24.00             | 66.76 ± 11.08            | 86.01 ± 12.42            |
| 5ms RTT         | 106.42 ± 6.01              | 35.93 ± 10.90            | 33.26 ± 7.11             |
| 10ms RTT        | 62.87 ± 12.44              | 21.19 ± 7.47             | 29.72 ± 2.00             |
| 25ms RTT        | 31.32 ± 2.10               | 24.45 ± 3.07             | 20.71 ± 8.40             |
| 50ms RTT        | 17.38 ± 0.46               | 14.51 ± 0.21             | 13.77 ± 0.41             |
| 100ms RTT       | 9.41 ± 0.09                | 7.41 ± 0.06              | 6.89 ± 0.54              |
+-----------------+----------------------------+--------------------------+--------------------------+
Based on 5 samples of 50 MB. All values in MB/s.
```

after:
```
+-----------------+----------------------------+--------------------------+--------------------------+
|                 | Chrome => TCP/HTTPS server | Chrome => quic-go server | quic-go client => server |
+-----------------+----------------------------+--------------------------+--------------------------+
| direct transfer | 125.48 ± 48.96             | 71.09 ± 9.13             | 90.69 ± 10.90            |
| 5ms RTT         | 98.09 ± 15.71              | 43.42 ± 6.92             | 41.92 ± 6.98             |
| 10ms RTT        | 75.95 ± 1.57               | 35.38 ± 5.07             | 40.23 ± 5.98             |
| 25ms RTT        | 33.01 ± 0.52               | 25.88 ± 1.27             | 24.09 ± 1.03             |
| 50ms RTT        | 16.75 ± 0.56               | 12.61 ± 3.36             | 12.96 ± 0.41             |
| 100ms RTT       | 8.48 ± 0.19                | 7.33 ± 0.26              | 7.07 ± 0.12              |
+-----------------+----------------------------+--------------------------+--------------------------+
Based on 5 samples of 50 MB. All values in MB/s.
```